+++ NEWS.md
@@ -1,3 +1,5 @@
+# fs (development version)
+
 # fs 1.3.2
 
 * fs now passes along `CPPFLAGS` during compilation of libuv, fixing an issue that could