+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * `path_ext_set()` now recycles both the `path` and `ext` arguments (#250).
 
+* Internally fs no longer depends on Rcpp
+
 # fs 1.3.2
 
 * fs now passes along `CPPFLAGS` during compilation of libuv, fixing an issue that could